server/requirements.txt
django>2
djangorestframework
djangorestframework-jwt
djangorestframework-stubs